+++ RPG Pirate - Roman Vitolo/Assets/Scripts/AgentLogic/AgentHealthSO.cs
@@ -5,7 +5,7 @@ namespace AgentLogic
     [CreateAssetMenu(menuName = "Entity/Agent Health Attributes", fileName = "New Agent Health")]
     public class AgentHealthSO : ScriptableObject
     {
-        [field: SerializeField] public float MaxAgentLife { get; set; } 
-        [field: SerializeField] public float CurrentAgentLife { get; set; } 
+        [field: SerializeField] public int MaxAgentLife { get; set; } 
+        [field: SerializeField] public int CurrentAgentLife { get; set; } 
     }
 }
